@@ -2,7 +2,7 @@ const TeamConstants = require('../constants/teams');
 const Player = require('./Player');
 
 class ServerState {
-  constructor (values) {
+  constructor(values) {
     this.live = false;
     this.map = '';
     this.maps = [];
@@ -29,35 +29,35 @@ class ServerState {
     this.players = {};
     this.pauses = {};
     this.last_log = undefined;
-    
+
     for (let key in values) {
       if (!this.hasOwnProperty(key) || !values.hasOwnProperty(key)) {
         continue;
       }
-      
+
       this[key] = values[key];
     }
     return this;
   }
-  
+
   /**
    * @returns {Player}
    */
-  getPlayer (steamId) {
+  getPlayer(steamId) {
     return this.players[steamId] ? this.players[steamId] : undefined;
   }
-  
-  addPlayer (steamId, team, name, clantag) {
+
+  addPlayer(steamId, team, name, clantag) {
     this.players[steamId] = new Player(steamId, team, name, clantag);
   }
-  
-  deletePlayer (steamId) {
+
+  deletePlayer(steamId) {
     if (this.players[steamId]) {
       delete this.players[steamId];
     }
   }
-  
-  clearPlayers () {
+
+  clearPlayers() {
     for (let steamId in this.players) {
       delete this.players[steamId];
     }
