@@ -1,5 +1,8 @@
-const SteamID = require('steamid');
-
-module.exports = steamid => {
-  return (new SteamID(String(steamid))).getSteamID64();
-};
+"use strict";
+exports.__esModule = true;
+var steamid_1 = require("steamid");
+var g = '';
+g = 123;
+exports["default"] = (function (steamid) {
+    return new steamid_1["default"](String(steamid)).getSteamID64();
+});
